Use readme from master and update gh-pages script

+++ gh-pages.sh
@@ -1,15 +1,21 @@
 #!/bin/bash
 
-git branch -D gh-pages
-git push origin --delete gh-pages
-git checkout -b gh-pages
-ember build --environment production
-git rm -rf app addon config tests
-git rm -rf Brocfile.js bower.json package.json testem.json
+ember build --prod
+git checkout gh-pages
+# from https://stackoverflow.com/questions/37890510/bash-script-to-check-if-the-current-git-branch-x
+BRANCH=$(git rev-parse --abbrev-ref HEAD)
+if [[ "$BRANCH" != "gh-pages" ]]; then
+  echo 'Aborting script';
+  exit 1;
+fi
+git rm -rf app addon config tests blueprints
+git rm -rf Brocfile.js bower.json package.json package-lock.json testem.json
 git rm -rf .bowerrc .editorconfig .jshintrc .travis.yml
-mv dist/* .
+git rm -rf v.1
+mkdir ./v.1
+mv dist/* ./v.1
 rm -rf dist
-git add .
-git commit -m "Publishing to github pages"
+git add v.1
+git commit -m "Pushing v.1 to github pages"
 git push origin gh-pages
 git checkout master
